+++ internal/web/static/app.js
@@ -18,7 +18,7 @@ import "./filehist.js";
 import "./review.js";
 import { applyStoredWidths } from "./resize.js";
 import { applyGraphMode, loadCommits, renderCommits } from "./commits.js";
-import { applyDiffView, applyFilesHidden, refreshNoteCounts } from "./files.js";
+import { applyDiffView, applyFilesHidden, applyTextMode, refreshNoteCounts } from "./files.js";
 import { focusPane } from "./keys.js";
 import { fetchHealth } from "./bigrepo.js";
 import "./settings.js";
@@ -58,6 +58,7 @@ async function applyStoredLayout() {
   if (ui.files_hidden) applyFilesHidden(true);
   applyGraphMode(ui.graph);
   applyDiffView(ui.diff_view);
+  applyTextMode(ui.text_mode);
 }